+++ src/contentScript/index.js
@@ -1,6 +1,6 @@
 import extensionizer from "extensionizer";
-import {sendMsg} from "../utils/commonMsg";
-import { MessageChannel,getSiteIcon } from '@aurowallet/mina-provider';
+import { sendMsg } from "../utils/commonMsg";
+import { MessageChannel, getSiteIcon } from "@aurowallet/mina-provider";
 import { WALLET_CONNECT_TYPE } from "../constant/commonType";
 
 const CONTENT_SCRIPT = WALLET_CONNECT_TYPE.CONTENT_SCRIPT;
@@ -8,16 +8,17 @@ const contentScript = {
   init() {
     this.channel = new MessageChannel(CONTENT_SCRIPT);
     this.registerListeners();
-    this.inject()
-    this.reportUrl()
+    this.inject();
+    this.reportUrl();
   },
-  reportUrl(){
+  reportUrl() {
     extensionizer.runtime.connect({ name: CONTENT_SCRIPT });
   },
   registerListeners() {
     window.addEventListener("message", (event) => {
-      const { data: eventData } = event;
+      const { data: eventData, isTrusted } = event;
       const { isAuro = false, message, source } = eventData;
+      if (!isTrusted) return;
       if (!isAuro || (!message && !source)) return;
       if (source === CONTENT_SCRIPT) return;
       const { data } = message;
@@ -29,13 +30,16 @@ const contentScript = {
           webIcon: getSiteIcon(event.source),
         },
       };
-      sendMsg({
-        action: data.action,
-        messageSource: 'messageFromDapp',
-        payload: nextPayload,
-      }, async (params) => {
-        this.channel.send("messageFromWallet", params);
-      });
+      sendMsg(
+        {
+          action: data.action,
+          messageSource: "messageFromDapp",
+          payload: nextPayload,
+        },
+        async (params) => {
+          this.channel.send("messageFromWallet", params);
+        }
+      );
     });
 
     extensionizer.runtime.onMessage.addListener(
@@ -52,14 +56,14 @@ const contentScript = {
   },
 
   inject() {
-    const hostPage = (document.head || document.documentElement);
-    const script = document.createElement('script');
+    const hostPage = document.head || document.documentElement;
+    const script = document.createElement("script");
 
-    script.src = extensionizer.runtime.getURL('webhook.js');
-    script.onload = function() {
+    script.src = extensionizer.runtime.getURL("webhook.js");
+    script.onload = function () {
       this.parentNode.removeChild(this);
     };
     hostPage.appendChild(script);
-  }
+  },
 };
 contentScript.init();